add "mass_offsets" parameter support (vectorMassOffsets variable)
git-svn-id: https://svn.code.sf.net/p/comet-ms/code@759 299d2d02-8f3b-485e-a07e-71e8e5e16577

+++ trunk/comet-ms/CometSearch/CometPostAnalysis.cpp
@@ -575,9 +575,9 @@ void CometPostAnalysis::LinearRegression(int *piHistogram,
 
    iStartCorr = 0;
    if (iNextCorr >= 30)
-      iStartCorr = iNextCorr - iNextCorr*0.25;
+      iStartCorr = (int)(iNextCorr - iNextCorr*0.25);
    else if (iNextCorr >= 15)
-      iStartCorr = iNextCorr - iNextCorr*0.5;
+      iStartCorr = (int)(iNextCorr - iNextCorr*0.5);
 
    Mx=My=a=b=0.0;
